Support Elixir 1.13
In version >= 1.13 some expressions previously without a line now have
a line propagated from a parent. This line does not always match
the location of the expression in the code. So we still need
to look at the tokens to correct these locations. This location presence
required small changes in the AstSpecifier code because now we can't be
sure when the expression has a line then this line is correct.

Changes:
- Adapt debug.ex to new elixir api
- Add specifying a line in nil.
- Propagate a generated flag from clauses to conditions.
- Consume the beginning token of list or tuple.
- Stop predicting the last line of the form from :bin or :cons expressions.